requirements.txt
pymongo==4.6.1
openai==1.6.1
python-dotenv==1.0.0
numpy==1.26.0
tqdm==4.66.1 